feat: use appEvents to handle email sending + move SMTP email logic to email plugin
