+++ imports/plugins/core/email/server/startup.js
@@ -1,0 +1,4 @@
+import appEvents from "/imports/node-app/core/util/appEvents";
+import sendSMTPEmail from "./util/sendSMTPEmail";
+
+appEvents.on("sendEmail", sendSMTPEmail);
